Performance: Refactored site history rendering
Site history queries were being issued for every new
site selection, and far too much data was being sent over
the wire to render on the browser.

We purposely do not want to commit to making all sitehistory
offline, so there's no reason we can't render it on the server.

Also an opportunity to start moving towards the new model
on the server, because it simplifies dealing with properties:
instead of three seperate special cases for projects, partners,
locations, etc, we treat them uniformally as reference fields.