Replaces List<> return types on RecordBeans with RecordArray<>
Implementing the List interface with JavaScript Overlay types
seems like it will cause a lot of problems, because it imposes
restrictions on how the List<> interface can be used across
the entire application.

This broke some random GXT code initially, for which we might be
able to rewrite/find a workaround but I can see running into this
repeatedly.

Not entirely convinced that the overlay approach is entirely
the best: even Backbone.Model essentially parses Json into
a JavaScript object...