Removes LocalizedString
This turned out not to be the way to implement what we want.
We want users to be able to provide translations of different structures,
such as FormClasses, etc.

But this seems to be better achieved by having a seperate set
of Translations that the user provides; easier to keep in sync,
less repeated translations.